removed redirects from OrdersList
